requirements.txt
Django>=1.8
factory-boy
